# Removed the tests of each layer with the sequential model.